requirements.txt
# AI Firewall POC - Requirements
# Python 3.8+

# Core ML Libraries
xgboost>=2.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# Data Processing
pyarrow>=12.0.0  # For Parquet support
fastparquet>=2023.7.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Configuration & Logging
pyyaml>=6.0
python-dotenv>=1.0.0

# Model Persistence
joblib>=1.3.0

# Performance
tqdm>=4.65.0

# Web API & Dashboard
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
python-multipart>=0.0.6

# System monitoring
psutil>=5.9.0

# Suricata/Zeek integration
watchdog>=3.0.0

# Redis
redis>=5.0.0

# Real-time packet capture
scapy>=2.5.0

# Optional: GPU Support (CUDA)
# Uncomment if using NVIDIA GPU with CUDA 11.8+
# xgboost[gpu]>=2.0.0
